feat: Moved update application code into applicationService.
